#127 [REFACTOR]: Elementos do tipo input agrupados em subdiretório
