+++ src/components/inputs/answers/TextImageInput.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import FormInputButtons from './FormInputButtons';
+import FormInputButtons from '../../FormInputButtons';
 
 const styles = `
     .font-barlow {
